server/pyproject.toml
[project]
name = "rag-chat-server"
version = "0.1.0"
description = "RAG Chat Server"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.3",
    "python-multipart>=0.0.9",
    "chromadb>=0.6.3",
    "beautifulsoup4>=4.13.3",
    "langchain-openai>=0.3.8",
    "langchain-community>=0.3.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
